Add base interface for objects with display properties.
This will allow the formatter to check if strings coming from
(for example) users in channels should be formatted specially,
such as by colouring or linking them.

Issues #424 and DMDirc/Plugins#507
